fastapi_yolo/requirements.txt
torch
torchvision
ultralytics
fastapi
uvicorn[standard]
aiofiles
Pillow
opencv-python-headless
